Accordionify dag details and fix tests
