+++ airflow/www/static/js/dag/details/dagRun/index.tsx
@@ -16,46 +16,21 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import React, { useEffect, useRef } from "react";
-import {
-  Flex,
-  Box,
-  Button,
-  Divider,
-  Spacer,
-  Table,
-  Tbody,
-  Tr,
-  Td,
-  useClipboard,
-} from "@chakra-ui/react";
-
-import ReactJson from "react-json-view";
-
+import React, { useRef } from "react";
+import { Box, Accordion } from "@chakra-ui/react";
 import { useGridData } from "src/api";
 import { getMetaValue, useOffsetTop } from "src/utils";
 import type { DagRun as DagRunType } from "src/types";
-import { SimpleStatus } from "src/dag/StatusBox";
-import { ClipboardText } from "src/components/Clipboard";
-import { formatDuration, getDuration } from "src/datetime_utils";
-import Time from "src/components/Time";
-import RunTypeIcon from "src/components/RunTypeIcon";
-import NotesAccordion from "src/dag/details/NotesAccordion";
+import Notes from "src/dag/details/Notes";
 
 import DatasetTriggerEvents from "./DatasetTriggerEvents";
+import Details from "./Details";
 
 const dagId = getMetaValue("dag_id");
 
 interface Props {
   runId: DagRunType["runId"];
 }
-
-const formatConf = (conf: string | null | undefined): string => {
-  if (!conf) {
-    return "";
-  }
-  return JSON.stringify(JSON.parse(conf), null, 4);
-};
 
 const DagRun = ({ runId }: Props) => {
   const {
@@ -65,27 +40,9 @@ const DagRun = ({ runId }: Props) => {
   const offsetTop = useOffsetTop(detailsRef);
 
   const run = dagRuns.find((dr) => dr.runId === runId);
-  const { onCopy, setValue, hasCopied } = useClipboard(formatConf(run?.conf));
-
-  useEffect(() => {
-    setValue(formatConf(run?.conf));
-  }, [run, setValue]);
 
   if (!run) return null;
-  const {
-    state,
-    runType,
-    lastSchedulingDecision,
-    dataIntervalStart,
-    dataIntervalEnd,
-    startDate,
-    endDate,
-    queuedAt,
-    externalTrigger,
-    conf,
-    confIsJson,
-    note,
-  } = run;
+  const { runType, note } = run;
 
   return (
     <Box
@@ -94,127 +51,18 @@ const DagRun = ({ runId }: Props) => {
       overflowY="auto"
       pb={4}
     >
-      <Box px={4}>
-        <NotesAccordion
+      <Accordion defaultIndex={[0]} allowMultiple>
+        <Details run={run} />
+        <Notes
           dagId={dagId}
           runId={runId}
           initialValue={note}
           key={dagId + runId}
         />
-      </Box>
-      <Divider my={0} />
-      <Table variant="striped">
-        <Tbody>
-          <Tr>
-            <Td>Status</Td>
-            <Td>
-              <Flex>
-                <SimpleStatus state={state} mx={2} />
-                {state || "no status"}
-              </Flex>
-            </Td>
-          </Tr>
-          <Tr>
-            <Td>Run ID</Td>
-            <Td>
-              <ClipboardText value={runId} />
-            </Td>
-          </Tr>
-          <Tr>
-            <Td>Run type</Td>
-            <Td>
-              <RunTypeIcon runType={runType} />
-              {runType}
-            </Td>
-          </Tr>
-          {startDate && (
-            <Tr>
-              <Td>Run duration</Td>
-              <Td>{formatDuration(getDuration(startDate, endDate))}</Td>
-            </Tr>
-          )}
-          {lastSchedulingDecision && (
-            <Tr>
-              <Td>Last scheduling decision</Td>
-              <Td>
-                <Time dateTime={lastSchedulingDecision} />
-              </Td>
-            </Tr>
-          )}
-          {queuedAt && (
-            <Tr>
-              <Td>Queued at</Td>
-              <Td>
-                <Time dateTime={queuedAt} />
-              </Td>
-            </Tr>
-          )}
-          {startDate && (
-            <Tr>
-              <Td>Started</Td>
-              <Td>
-                <Time dateTime={startDate} />
-              </Td>
-            </Tr>
-          )}
-          {endDate && (
-            <Tr>
-              <Td>Ended</Td>
-              <Td>
-                <Time dateTime={endDate} />
-              </Td>
-            </Tr>
-          )}
-          {dataIntervalStart && dataIntervalEnd && (
-            <>
-              <Tr>
-                <Td>Data interval start</Td>
-                <Td>
-                  <Time dateTime={dataIntervalStart} />
-                </Td>
-              </Tr>
-              <Tr>
-                <Td>Data interval end</Td>
-                <Td>
-                  <Time dateTime={dataIntervalEnd} />
-                </Td>
-              </Tr>
-            </>
-          )}
-          <Tr>
-            <Td>Externally triggered</Td>
-            <Td>{externalTrigger ? "True" : "False"}</Td>
-          </Tr>
-          <Tr>
-            <Td>Run config</Td>
-            {confIsJson ? (
-              <Td>
-                <Flex>
-                  <ReactJson
-                    src={JSON.parse(conf ?? "")}
-                    name={false}
-                    theme="rjv-default"
-                    iconStyle="triangle"
-                    indentWidth={2}
-                    displayDataTypes={false}
-                    enableClipboard={false}
-                    style={{ backgroundColor: "inherit" }}
-                  />
-                  <Spacer />
-                  <Button aria-label="Copy" onClick={onCopy}>
-                    {hasCopied ? "Copied!" : "Copy"}
-                  </Button>
-                </Flex>
-              </Td>
-            ) : (
-              <Td>{conf ?? "None"}</Td>
-            )}
-          </Tr>
-        </Tbody>
-      </Table>
-      {runType === "dataset_triggered" && (
-        <DatasetTriggerEvents runId={runId} />
-      )}
+        {runType === "dataset_triggered" && (
+          <DatasetTriggerEvents runId={runId} />
+        )}
+      </Accordion>
     </Box>
   );
 };
